Add files via upload
The SMS Spam Collection Dataset is utilized. The dataset contains the corpus of 5572 SMS messages and a label, classifying the message as "spam" or "ham" (legitimate). An automated  spam detector is built  with various classifiers and finally a Data Pipeline is set up to store a pipeline of workflow for future predictions using a specific model.